Refactored settingsutils
